fix: tests

+++ pgdump.go
@@ -261,7 +261,7 @@ func parseTypeOption(tp sqlast.SQLType, info *pgInformationSchemaColumns) sqlast
 	case *sqlast.VarcharType:
 		if info.CharacterMaximumLength.Valid {
 			return &sqlast.VarcharType{
-				Size: sqlast.NewSize(uint8(info.CharacterMaximumLength.Int64)),
+				Size: sqlast.NewSize(uint(info.CharacterMaximumLength.Int64)),
 			}
 		}
 		return tp
